Add port binding as flow reference
We have mechanism to track flows updating address set and port group
and then apply the tracked flows when these resources updated.
Similarly we track flows related to ports and in port binding
handler we apply these flows for local port binding changes.

We add port_binding as new engine node for tracking its changes,
as _flow_output_resource_ref_handler expects engine node to have
"new", "deleted" and "updated" changess as part of engine data.

When port binding updated on the local host, it port groups and ACLs
have to be applied. So we identify the port groups to which this
port belongs and then run _flow_output_resource_ref_handler on
these port groups.

When a port is added to host, if this is first port on that host
on that local datapath or if it is last port on that host on that
local datapath, then we need to identify that and apply flows
related to local datapth. For now, we trigger engine force recompute
as we are not tracking flows for the datapath (can be enhanced later).

We cache local ports and local datapaths from previous run and
then compare with current local port and datapaths (which are rebuilt
in runtime_data_sb_port_binding_handler) to identify if a port is
added to this host or deleted from this host along with port binding
data.