[WIP] optimize port groups and address sets
ovn-controller recalculates flows for all the members of a PG or AS
when a new member is added or deleted. Instead, this patch identifies
the changed members and only process their related flows.

Also this patch maintains constant sets for only local ports of a port
group (instead of all ports) to facilitate lesser lexer parsing and
symbols in expr.

This patch is unable to use conjuction flows (as conjuction ids are not
stored) and generate below flows
priority=2002,ct_state=-trk,ip,reg15=0x2,metadata=0x2,nw_src=10.0.0.239 actions=resubmit(,45)

When conjuction ids used by a logical flow are maintained as part of
flow data structures instead of dynamically regenerated during flow
processing, we can update the patch to use the related conjuction ids
for the above flows.

Will test performance impact once the patch is updated with conjuction
flows.